+++ python/src/Python.cpp
@@ -3,6 +3,7 @@
 #include <Jet.hpp>
 
 #include "PathInfo.hpp"
+#include "TaskBasedCpuContractor.hpp"
 #include "Tensor.hpp"
 #include "TensorNetwork.hpp"
 #include "TensorNetworkIO.hpp"
@@ -16,6 +17,9 @@ PYBIND11_MODULE(bindings, m)
     using c128_t = std::complex<double>;
 
     AddBindingsForPathInfo<c64_t, c128_t>(m);
+
+    AddBindingsForTaskBasedCpuContractor<c64_t>(m);
+    AddBindingsForTaskBasedCpuContractor<c128_t>(m);
 
     AddBindingsForTensor<c64_t>(m);
     AddBindingsForTensor<c128_t>(m);