+++ python/tests/jet/test_tensor_network_io.py
@@ -16,22 +16,6 @@ def test_tensor_network_file(dtype):
 @pytest.mark.parametrize("dtype", ["complex64", "complex128"])
 class TestTensorNetworkSerializer:
     @pytest.fixture
-    def tensor_network(self):
-        def tensor_network_(dtype: str):
-            """Returns a tensor network with three tensors of the given type."""
-            A = jet.Tensor(shape=[2, 2], indices=["i", "j"], data=[1, 1j, -1j, 1], dtype=dtype)
-            B = jet.Tensor(shape=[2, 2], indices=["j", "k"], data=[1, 0, 0, 1], dtype=dtype)
-            C = jet.Tensor(shape=[2], indices=["k"], data=[1, 0], dtype=dtype)
-
-            tn = jet.TensorNetwork(dtype=dtype)
-            tn.add_tensor(A, ["A", "Hermitian"])
-            tn.add_tensor(B, ["B", "Identity", "Real"])
-            tn.add_tensor(C, ["C", "Vector", "Real"])
-            return tn
-
-        return tensor_network_
-
-    @pytest.fixture
     def serialized_tensor_network(self) -> str:
         """Returns a serialized tensor network file representing a tensor
         network.